requirement.txt
bcrypt==2.0.0
django-crispy-forms==1.6.0
django-registration-redux==1.4

